--- src/beignet/func/_molecular_dynamics/_interact/__angle_interaction.py
@@ -1,10 +1,0 @@
-from typing import Callable
-
-from torch import Tensor
-
-
-def _angle_interaction(
-    fn: Callable[..., Tensor],
-    displacement_fn: Callable[[Tensor, Tensor], Tensor],
-):
-    raise NotImplementedError
