--- src/beignet/func/_molecular_dynamics/_interact/__parameter_tree_kind.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class _ParameterTreeKind(Enum):
-    BOND = 0
-    KINDS = 1
-    PARTICLE = 2
-    SPACE = 3
